move RewardsInfo to dedicated file
